requirements.txt
flask
requests
textblob
python-dotenv